Rework code and Makefile to use m4
I want to be able to define where the 'functions' directory and its code
is, but the binaries and the functions code will probably live in
different directories. I think the right thing to do is to use a
preprocessor to set the file path as part of an install process, and
test off of that. So here I'm using m4 to preprocess the files.

Functions directory moving to the parent (root) directory, so that it
can be used by frontend and other code.